+++ engine/flink/management/dev-model/src/test/scala/pl/touk/nussknacker/engine/process/NamespacedKafkaSourceSinkTest.scala
@@ -46,7 +46,7 @@ class NamespacedKafkaSourceSinkTest extends AnyFunSuite with FlinkSpec with Kafk
       .emptySink("output", "kafka-string", TopicParamName -> s"'$outputTopic'", SinkValueParamName -> "#input")
 
     run(process) {
-      val processed = kafkaClient.createConsumer().consume[String](s"ns_$outputTopic", 1).map(_.message())
+      val processed = kafkaClient.createConsumer().consumeWithJson[String](s"ns_$outputTopic", 1).map(_.message())
       processed shouldEqual List(message)
     }
   }