+++ engine/lite/embeddedDeploymentManager/src/test/scala/pl/touk/nussknacker/streaming/embedded/StreamingEmbeddedDeploymentManagerTest.scala
@@ -47,7 +47,7 @@ class StreamingEmbeddedDeploymentManagerTest extends BaseStreamingEmbeddedDeploy
 
     val input = obj("productId" -> fromInt(10))
     kafkaClient.sendMessage(inputTopic, input.noSpaces).futureValue
-    kafkaClient.createConsumer().consume[Json](outputTopic).take(1).head.message() shouldBe input
+    kafkaClient.createConsumer().consumeWithJson[Json](outputTopic).take(1).head.message() shouldBe input
 
     wrapInFailingLoader {
       manager.cancel(name, User("a", "b")).futureValue
@@ -75,7 +75,7 @@ class StreamingEmbeddedDeploymentManagerTest extends BaseStreamingEmbeddedDeploy
     val input = obj("productId" -> fromInt(10))
     kafkaClient.sendMessage(inputTopic, input.noSpaces).futureValue
 
-    kafkaClient.createConsumer().consume[Json](outputTopic).take(1).head.message() shouldBe input
+    kafkaClient.createConsumer().consumeWithJson[Json](outputTopic).take(1).head.message() shouldBe input
 
     manager.cancel(name, User("a", "b")).futureValue
     manager.getProcessStates(name).futureValue.value shouldBe List.empty
@@ -146,7 +146,9 @@ class StreamingEmbeddedDeploymentManagerTest extends BaseStreamingEmbeddedDeploy
     fixture.deployScenario(scenarioForOutput("start"))
 
     kafkaClient.sendMessage(inputTopic, message("1")).futureValue
-    kafkaClient.createConsumer().consume[Json](outputTopic).take(1).head.message() shouldBe prefixMessage("start", "1")
+
+    val consumer = kafkaClient.createConsumer().consumeWithJson[Json](outputTopic).map(_.message())
+    consumer.head shouldBe prefixMessage("start", "1")
 
     fixture.deployScenario(scenarioForOutput("next"))
 
@@ -155,15 +157,11 @@ class StreamingEmbeddedDeploymentManagerTest extends BaseStreamingEmbeddedDeploy
     }
 
     kafkaClient.sendMessage(inputTopic, message("2")).futureValue
-    kafkaClient.createConsumer().consume[Json](outputTopic).take(2).map(_.message()) shouldBe List(
-      prefixMessage("start", "1"), prefixMessage("next", "2")
-    )
+    consumer.take(2) shouldBe List(prefixMessage("start", "1"), prefixMessage("next", "2"))
 
     kafkaClient.sendMessage(inputTopic, message("3")).futureValue
-    kafkaClient.createConsumer().consume[Json](outputTopic).take(3).map(_.message()) shouldBe List(
-      prefixMessage("start", "1"), prefixMessage("next" , "2"),
-      prefixMessage("next", "3")
-    )
+    consumer.take(3) shouldBe List(prefixMessage("start", "1"), prefixMessage("next" , "2"),
+      prefixMessage("next", "3"))
 
     manager.cancel(name, User("a", "b")).futureValue