+++ engine/lite/integration-test/src/it/scala/pl/touk/nussknacker/engine/lite/kafka/NuKafkaRuntimeDockerJsonTest.scala
@@ -29,7 +29,7 @@ class NuKafkaRuntimeDockerJsonTest extends AnyFunSuite with BaseNuKafkaRuntimeDo
   test("json ping-pong should work") {
     kafkaClient.sendMessage(fixture.inputTopic, NuKafkaRuntimeTestSamples.jsonPingMessage).futureValue
     try {
-      val message = kafkaClient.createConsumer().consume[String](fixture.outputTopic).head.message()
+      val message = kafkaClient.createConsumer().consumeWithJson[String](fixture.outputTopic).head.message()
       message shouldBe NuKafkaRuntimeTestSamples.jsonPingMessage
     } finally {
       consumeFirstError shouldBe empty